% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interana_data.R
\name{interana_add_query_params}
\alias{interana_add_query_params}
\title{Add Interana Query Params}
\usage{
interana_add_query_params(query_type = "single_measurement",
  agg_type = "count_star", agg_column = "", filter_expr = "",
  group_by_column = "")
}
\arguments{
\item{query_type}{Specify the Interana query type. Default=single_measurement. More details at TBD}

\item{agg_type}{Specify the Interana aggregator type. Default=count_star. More details at TBD}

\item{agg_column}{Specify the aggregator column, if any. This is dependent on the aggregator type. Default=null. More details at TBD}

\item{filter_expr}{Specify the filter criteria, if any. This is an URL-escaped value (example: TBD) Default=null. More details at TBD}

\item{group_by_column}{Specify the group by column. Default=null. More details at TBD}
}
\description{
This function allows you to further specify Interana Query Params
}
\examples{
interana_add_query_params()
}
\keyword{interana}
