% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interana_data.R
\name{interana_client}
\alias{interana_client}
\title{Establish the Interana Client Connection}
\usage{
interana_client(cluster_host = "", token = "")
}
\arguments{
\item{cluster_host}{Specify the cluster domain. For eg. https://<cluster-domain>/login.html. This is usually the name of your company or POC with a suffix of .interana.com.}

\item{token}{Provide the authorization token to validate access to the cluster above.}
}
\description{
This function allows you to establish the Interana Client Connection
}
\examples{
interana_client()
}
\keyword{interana}
