% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interana_data.R
\name{interana_get_data}
\alias{interana_get_data}
\title{Retrieve the Interana Results from previously formulated data}
\usage{
interana_get_data()
}
\description{
This function allows you to retrieve data from previously defined query
}
\examples{
interana_get_data()
}
\keyword{interana}
