% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interana_data.R
\name{interana_get_params}
\alias{interana_get_params}
\title{Gets the parameters that are set so far}
\usage{
interana_get_params()
}
\description{
This function shows all the parameters set so far.
}
\examples{
interana_get_params()
}
\keyword{get_params}
