% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interana_data.R
\name{interana_query}
\alias{interana_query}
\title{Formulate the Interana Query}
\usage{
interana_query(dataset = "", days_prior = 7)
}
\arguments{
\item{dataset}{Specify the Interana dataset to query against}

\item{days_prior}{Specify the interval to search against. In other words: end_time: now, start_time: now-days_prior}
}
\description{
This function allows you to define the Interana Query
}
\examples{
interana_query()
}
\keyword{interana}
